feat(migration-guide): add migration guide for AI SDK 4.0 to 5.0 Beta to assist developers in upgrading their projects fix(schema): update Message model to remove deprecated fields and enforce required parts field for better data integrity fix(chat): refactor chat components to support new message structure and improve user experience with better error handling and message management chore(dependencies): update package versions to ensure compatibility with AI SDK 5.0 Beta and improve overall application performance style(chat): enhance UI components for better visual consistency and user interaction refactor(chat): streamline chat logic and state management for improved maintainability and performance test(chat): add tests for new chat functionalities to ensure reliability and correctness after migration to AI SDK 5.0 Beta
feat(chat): implement chat component with message handling and UI features to enhance user interaction
feat(chat): create message list component to display chat messages and typing indicators
feat(chat): add prompt suggestions component to assist users with message input
feat(chat): establish message reducer for managing message state and actions
feat(chat): introduce chat form for user input and message submission
chore(chat): restructure chat-related components into a dedicated directory for better organization
refactor(chat): remove deprecated message list component and related schemas for cleaner codebase
refactor(schema): update message schemas to include metadata and improve compatibility with the database
refactor(message.ts): streamline message router methods for better readability and maintainability
