+++ adaptive-app/package.json
@@ -19,8 +19,9 @@
 		"stripe": "stripe listen --forward-to localhost:3000/api/stripe-checkout"
 	},
 	"dependencies": {
-		"@ai-sdk/openai": "^1.3.22",
-		"@clerk/nextjs": "^6.22.0",
+		"@ai-sdk/openai": "^2.0.0-beta.2",
+		"@ai-sdk/react": "^2.0.0-beta.3",
+		"@clerk/nextjs": "^6.23.1",
 		"@hookform/resolvers": "^5.1.1",
 		"@prisma/client": "^6.10.1",
 		"@radix-ui/react-avatar": "^1.1.10",
@@ -36,32 +37,33 @@
 		"@radix-ui/react-switch": "^1.2.5",
 		"@radix-ui/react-tabs": "^1.1.12",
 		"@radix-ui/react-tooltip": "^1.2.7",
-		"@shikijs/transformers": "^3.6.0",
-		"@t3-oss/env-nextjs": "^0.12.0",
-		"@tanstack/react-query": "^5.80.7",
-		"@trpc/client": "^11.4.2",
-		"@trpc/react-query": "^11.4.2",
-		"@trpc/server": "^11.4.2",
+		"@shikijs/transformers": "^3.7.0",
+		"@t3-oss/env-nextjs": "^0.13.8",
+		"@tanstack/react-query": "^5.81.5",
+		"@trpc/client": "^11.4.3",
+		"@trpc/react-query": "^11.4.3",
+		"@trpc/server": "^11.4.3",
 		"@vercel/analytics": "^1.5.0",
-		"ai": "^4.3.16",
+		"ai": "^5.0.0-beta.3",
 		"class-variance-authority": "^0.7.1",
 		"clsx": "^2.1.1",
 		"cobe": "^0.6.4",
 		"date-fns": "^4.1.0",
-		"framer-motion": "^12.18.1",
-		"lucide-react": "^0.513.0",
-		"next": "^15.3.3",
+		"framer-motion": "^12.19.2",
+		"install": "^0.13.0",
+		"lucide-react": "^0.525.0",
+		"next": "^15.3.4",
 		"next-themes": "^0.4.6",
 		"react": "^19.1.0",
 		"react-dom": "^19.1.0",
-		"react-hook-form": "^7.58.1",
+		"react-hook-form": "^7.59.0",
 		"react-intersection-observer": "^9.16.0",
 		"react-markdown": "^10.1.0",
-		"recharts": "^2.15.4",
+		"recharts": "^3.0.2",
 		"remark-gfm": "^4.0.1",
-		"remeda": "^2.23.0",
+		"remeda": "^2.23.2",
 		"server-only": "^0.0.1",
-		"shiki": "^1.29.2",
+		"shiki": "^3.7.0",
 		"sonner": "^2.0.5",
 		"stripe": "^18.2.1",
 		"superjson": "^2.2.2",
@@ -70,14 +72,14 @@
 		"zod": "^3.25.67"
 	},
 	"devDependencies": {
-		"@biomejs/biome": "1.9.4",
-		"@tailwindcss/postcss": "^4.1.10",
-		"@types/node": "^20.19.1",
+		"@biomejs/biome": "^2.0.6",
+		"@tailwindcss/postcss": "^4.1.11",
+		"@types/node": "^24.0.7",
 		"@types/react": "^19.1.8",
 		"@types/react-dom": "^19.1.6",
 		"postcss": "^8.5.6",
 		"prisma": "^6.10.1",
-		"tailwindcss": "^4.1.10",
+		"tailwindcss": "^4.1.11",
 		"tw-animate-css": "^1.3.4",
 		"typescript": "^5.8.3"
 	},
