+++ adaptive-app/src/app/_components/chat-platform/chats/chat-client.tsx
@@ -1,115 +1,106 @@
 "use client";
 
+import { useCallback, useState } from "react";
+import { useChat, type UIMessage } from "@ai-sdk/react";
+import { DefaultChatTransport } from "ai";
+
 import { Chat } from "@/components/ui/chat";
 import { useMessageLimits } from "@/hooks/messages/use-message-limits";
 import type { ConversationListItem, Message } from "@/types";
-import { useChat } from "ai/react";
-import type { Message as SDKMessage } from "ai/react";
-import { useCallback, useMemo } from "react";
+
+const CHAT_SUGGESTIONS = [
+  "Generate a tasty vegan lasagna recipe for 3 people.",
+  "Generate a list of 5 questions for a frontend job interview.",
+  "Who won the 2022 FIFA World Cup?",
+];
 
 interface ChatClientProps {
-	conversation: ConversationListItem;
-	initialMessages: Message[];
+  conversation: ConversationListItem;
+  initialMessages: Message[];
 }
 
 export function ChatClient({ conversation, initialMessages }: ChatClientProps) {
-	const {
-		isUnlimited,
-		remainingMessages,
-		hasReachedLimit,
-		isLoading: limitsLoading,
-	} = useMessageLimits();
+  const {
+    isUnlimited,
+    remainingMessages,
+    hasReachedLimit,
+    isLoading: limitsLoading,
+  } = useMessageLimits();
 
-	const mappedMessages = useMemo(
-		() =>
-			initialMessages.map((msg: Message) => ({
-				id: msg.id,
-				role: msg.role as "user" | "assistant" | "system",
-				content: msg.content,
-				createdAt: msg.createdAt,
-				reasoning: msg.reasoning ?? undefined,
-				annotations: msg.annotations ? msg.annotations : undefined,
-				parts: msg.parts ? msg.parts : undefined,
-				experimentalAttachments: msg.experimentalAttachments
-					? msg.experimentalAttachments
-					: undefined,
-			})) as unknown as SDKMessage[],
-		[initialMessages],
-	);
+  const mappedMessages = initialMessages as unknown as UIMessage[];
 
-	const {
-		messages,
-		input,
-		handleInputChange,
-		handleSubmit: originalHandleSubmit,
-		setMessages,
-		append,
-		status,
-		stop,
-		error,
-		reload,
-	} = useChat({
-		api: "/api/chat",
-		id: conversation.id.toString(),
-		initialMessages: mappedMessages,
-		experimental_prepareRequestBody({ messages, id }) {
-			return { message: messages[messages.length - 1], id };
-		},
-	});
+  const [input, setInput] = useState("");
+  const {
+    messages,
+    setMessages,
+    status,
+    stop,
+    error,
+    sendMessage,
+    regenerate,
+  } = useChat({
+    id: conversation.id.toString(),
+    messages: mappedMessages,
+    transport: new DefaultChatTransport({
+      api: "/api/chat",
+      credentials: "include",
+    }),
+  });
 
-	const handleSubmit = useCallback(
-		(
-			event?: { preventDefault?: () => void },
-			options?: { experimental_attachments?: FileList },
-		) => {
-			// Call preventDefault if event exists and has the method
-			if (event?.preventDefault) {
-				event.preventDefault();
-			}
+  const handleInputChange = useCallback(
+    (event: React.ChangeEvent<HTMLTextAreaElement>) => {
+      setInput(event.target.value);
+    },
+    [],
+  );
 
-			if (hasReachedLimit) {
-				alert(
-					"You've reached your daily message limit. Please upgrade to continue.",
-				);
-				return;
-			}
+  const handleSubmit = useCallback(
+    (
+      event?: { preventDefault?: () => void },
+      options?: { files?: FileList },
+    ) => {
+      event?.preventDefault?.();
 
-			// Pass both parameters to the original function
-			originalHandleSubmit(event, options);
-		},
-		[originalHandleSubmit, hasReachedLimit],
-	);
+      if (hasReachedLimit) {
+        alert(
+          "You've reached your daily message limit. Please upgrade to continue.",
+        );
+        return;
+      }
 
-	const isLoading = status === "streaming" || status === "submitted";
-	const isError = status === "error";
+      if (!input.trim()) return;
 
-	return (
-		<div className="flex h-full flex-col">
-			<Chat
-				className="flex-1"
-				messages={messages}
-				input={input}
-				handleInputChange={handleInputChange}
-				handleSubmit={handleSubmit}
-				setMessages={setMessages}
-				isGenerating={isLoading}
-				stop={stop}
-				append={append}
-				suggestions={[
-					"Generate a tasty vegan lasagna recipe for 3 people.",
-					"Generate a list of 5 questions for a frontend job interview.",
-					"Who won the 2022 FIFA World Cup?",
-				]}
-				isError={isError}
-				error={error}
-				onRetry={reload}
-				// Pass limit info to Chat component
-				hasReachedLimit={hasReachedLimit}
-				remainingMessages={remainingMessages}
-				isUnlimited={isUnlimited}
-				limitsLoading={limitsLoading}
-				userId={conversation.userId} // Add this prop with the actual user ID
-			/>
-		</div>
-	);
+      sendMessage({ text: input, files: options?.files });
+      setInput("");
+    },
+    [sendMessage, hasReachedLimit, input],
+  );
+
+  const isLoading = status === "streaming" || status === "submitted";
+  const isError = status === "error";
+
+  return (
+    <div className="flex h-full flex-col">
+      <Chat
+        className="flex-1"
+        messages={messages}
+        input={input}
+        handleInputChange={handleInputChange}
+        handleSubmit={handleSubmit}
+        setMessages={setMessages}
+        isGenerating={isLoading}
+        stop={stop}
+        sendMessage={(message) => sendMessage({ text: message.text })}
+        suggestions={CHAT_SUGGESTIONS as string[]}
+        isError={isError}
+        error={error}
+        onRetry={regenerate}
+        hasReachedLimit={hasReachedLimit}
+        remainingMessages={remainingMessages}
+        isUnlimited={isUnlimited}
+        limitsLoading={limitsLoading}
+        userId={conversation.userId}
+      />
+    </div>
+  );
 }
