+++ adaptive-app/src/components/ui/chat/chat-container.tsx
@@ -1,0 +1,16 @@
+import { forwardRef } from "react";
+import { cn } from "@/lib/utils";
+
+export const ChatContainer = forwardRef<
+  HTMLDivElement,
+  React.HTMLAttributes<HTMLDivElement>
+>(({ className, ...props }, ref) => {
+  return (
+    <div
+      ref={ref}
+      className={cn("grid max-h-full w-full grid-rows-[1fr_auto]", className)}
+      {...props}
+    />
+  );
+});
+ChatContainer.displayName = "ChatContainer";
