+++ adaptive-app/src/components/ui/chat/index.ts
@@ -1,0 +1,9 @@
+export { Chat } from "./chat";
+export { ChatContainer } from "./chat-container";
+export { ChatErrorDisplay } from "./chat-error-display";
+export { ChatForm } from "./chat-form";
+export { ChatMessages } from "./chat-messages";
+export { useOptimisticMessageCount } from "./chat-hooks";
+export { messageReducer } from "./chat-reducer";
+export * from "./chat-types";
+export * from "./chat-utils";
