+++ adaptive-app/src/components/ui/chat/prompt-suggestions.tsx
@@ -1,12 +1,12 @@
 interface PromptSuggestionsProps {
 	label: string;
-	append: (message: { role: "user"; content: string }) => void;
+	sendMessage: (message: { text: string }) => void;
 	suggestions: string[];
 }
 
 export function PromptSuggestions({
 	label,
-	append,
+	sendMessage,
 	suggestions,
 }: PromptSuggestionsProps) {
 	return (
@@ -17,7 +17,7 @@ export function PromptSuggestions({
 					<button
 						type="button"
 						key={suggestion}
-						onClick={() => append({ role: "user", content: suggestion })}
+						onClick={() => sendMessage({ text: suggestion })}
 						className="h-max flex-1 rounded-xl border bg-background p-4 hover:bg-muted"
 					>
 						<p>{suggestion}</p>
